Add cuda version

+++ src/nbody.cpp
@@ -15,7 +15,7 @@ void NBodyEnv::run() {
     while (!displayer.stop()) {
         if (displayer.getRestart()) {
             displayer.setRestart(false);
-            for (int i = 0 ; i < particles.size() ; i++)
+            for (int i = 0 ; i < (int)particles.size() ; i++)
                 particles[i] = Particle(particles[i].mass);
         }
         displayer.update(particles);
